Split plugins._base into ._base and ._registry
