+++ libs/stats/odc/stats/plugins/_base.py
@@ -1,11 +1,10 @@
-import pydoc
 from abc import ABC, abstractmethod
-from typing import Callable, Dict, Tuple, Sequence, Optional
-from functools import partial
+from typing import Tuple, Sequence, Optional
 
 import xarray as xr
 from datacube.model import Dataset
 from datacube.utils.geometry import GeoBox
+
 
 class StatsPluginInterface(ABC):
     NAME = "*unset*"
@@ -31,52 +30,3 @@ class StatsPluginInterface(ABC):
         Given result of ``.reduce(..)`` optionally produce RGBA preview image
         """
         return None
-
-
-PluginFactory = Callable[..., StatsPluginInterface]
-
-_plugins: Dict[str, PluginFactory] = {}
-
-
-def _new(plugin_class, *args, **kwargs) -> StatsPluginInterface:
-    return plugin_class(*args, **kwargs)
-
-
-def resolve(name: str) -> PluginFactory:
-    maker = _plugins.get(name)
-    if maker is None:
-        maker = pydoc.locate(name)
-        if maker is not None:
-            if not issubclass(maker, (StatsPluginInterface,)):
-                raise ValueError(f"Custom StatsPlugin '{name}' is not derived from StatsPluginInterface")
-            return partial(_new, maker)
-
-    if maker is None:
-        raise ValueError(f"Failed to resolved named plugin: '{name}'")
-    return maker
-
-
-def register(name: str, plugin_class):
-    _plugins[name] = partial(_new, plugin_class)
-
-
-def import_all():
-    import importlib
-
-    # TODO: make that more automatic
-    modules = [
-        "odc.stats.plugins.fc_percentiles",
-        "odc.stats.plugins.tcw_percentiles",
-        "odc.stats.plugins.gm",
-        "odc.stats.plugins.gm_ls_bitmask",
-        "odc.stats.plugins.pq",
-        "odc.stats.plugins.pq_bitmask",
-        "odc.stats.plugins.wofs",
-    ]
-
-    for mod in modules:
-        try:
-            importlib.import_module(mod)
-        except ModuleNotFoundError:
-            pass
-
